Fixed %x and %X

+++ ft_printf_x.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_printf_%x.c                                     :+:      :+:    :+:   */
+/*   ft_printf_x.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jschneid <jschneid@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/30 16:10:38 by jschneid          #+#    #+#             */
-/*   Updated: 2022/05/31 09:15:17 by jschneid         ###   ########.fr       */
+/*   Updated: 2022/05/31 11:28:17 by jschneid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,10 +17,10 @@
 // and returns the number of digits wich are printed
 int	output_lowercase_hexa(va_list arguments)
 {
-	int					length_address;
-	int					ptr;
+	long long	length_address;
+	long long	ptr;
 
-	ptr = va_arg(arguments, int);
+	ptr = va_arg(arguments, long long);
 	length_address = get_number_length(ptr);
 	print_lowercase_hexa(ptr);
 	return (length_address);
@@ -56,10 +56,10 @@ void	print_lowercase_hexa(unsigned long long decimal_nbr)
 // and returns the number of digits wich are printed
 int	output_uppercase_hexa(va_list arguments)
 {
-	int					length_address;
-	int					ptr;
+	long long	length_address;
+	long long	ptr;
 
-	ptr = va_arg(arguments, int);
+	ptr = va_arg(arguments, long long);
 	length_address = get_number_length(ptr);
 	print_uppercase_hexa(ptr);
 	return (length_address);
